TS: various minor improvements.
