+++ src/main/ApplicationWindow.ts
@@ -2,7 +2,7 @@ import electron from 'electron'
 import path from 'path'
 
 export class ApplicationWindow extends electron.BrowserWindow {
-  alreadyShowedWindow = false
+  #alreadyShowedWindow = false
 
   constructor(options: electron.BrowserWindowConstructorOptions) {
     // Add some common options and call our parent constructor.
@@ -25,8 +25,8 @@ export class ApplicationWindow extends electron.BrowserWindow {
     //       multiple times (e.g., when returning from sleep mode on macOS).
 
     this.on('ready-to-show', () => {
-      if (!this.alreadyShowedWindow) {
-        this.alreadyShowedWindow = true
+      if (!this.#alreadyShowedWindow) {
+        this.#alreadyShowedWindow = true
 
         this.show()
       }
