+++ src/preload/index.ts
@@ -26,10 +26,10 @@ electron.contextBridge.exposeInMainWorld('electronAPI', {
 
   // Renderer process asking the main process to do something for it.
 
-  disableMenu: () => electron.ipcRenderer.invoke('disable-menu'),
-  enableMenu: () => electron.ipcRenderer.invoke('enable-menu'),
   filePath: (file: File) => electron.webUtils.getPathForFile(file),
   resetAll: () => electron.ipcRenderer.invoke('reset-all'),
+  showDisabledMenu: () => electron.ipcRenderer.invoke('show-disabled-menu'),
+  showEnabledMenu: () => electron.ipcRenderer.invoke('show-enabled-menu'),
 
   // Renderer process listening to the main process.
